app/requirements.txt
Flask>=2.0.2
PyJWT>=2.6.0